Added some styling to body. Also refactored each prompt into its own component
