+++ client/source/styles/style.css
@@ -1,3 +1,4 @@
 body {
   margin: 0;
+  background-color: #cccccc;
 }
