+++ package.json
@@ -39,6 +39,7 @@
     "material-ui": "^0.15.0",
     "mongoose": "^4.4.19",
     "react": "^15.1.0",
+    "react-addons-css-transition-group": "^15.1.0",
     "react-dom": "^15.1.0",
     "react-tap-event-plugin": "^1.0.0"
   },
